Use a u8 instead of int for Port::state. This allows better struct packing and reduces the size of objects by 4 bytes.

+++ portlist.h
@@ -182,7 +182,7 @@ class Port {
 
   u16 portno;
   u8 proto;
-  int state; 
+  u8 state; 
   state_reason_t reason;
 
 #ifndef NOLUA
